// src/web/src/api/generated/models/ExerciseMediaResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ExerciseMediaResponse = {
    id: string;
    position: number | string;
    assetUri: string;
    altText: string;
    mimeType: string;
    source: string;
    author: string | null;
    license: string;
    sha256: string | null;
    presentationSex: string;
    widthPx: number | string;
    heightPx: number | string;
};